DERBY-4743 Attempt to access a DECLARE GLOBAL TEMPORARY TABLE after commit in an XA transaction in Network server fails.
The fix for DERBY-4731 made the behavior consistent across embedded and 
network server.  Derby does not support accessing global temporary tables 
declared in an XA transaction subsequent to the commit of the transaction.

This change updates the junit XATest to verify that accessing the table after
the commit fails as expected in both the embedded and network server case.  In
both cases you will get the generic table does not exist error.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@966027 13f79535-47bb-0310-9956-ffa450edef68