DERBY-4746: Server support for UTF8 in DRDA
This patch was contributed by Tiago R. Espinha (tiago dot derby at yahoo dot co dot uk)

The code portion of this change puts in place all the switches necessary
on the server-side to enable and disable UTF8 support. Principally, the
new internal methods switchToUtf8() and switchToEbcdic() are added to
DRDAConnThread, and called appropriately from the ACCSEC code point.

The test portion of this change contains the protocol tests and changes
to the framework required for the UTF8 protocol tests. The test changes
implement the command 'switchToUtf8CcsidManager' in the protocol tests
that essentially tells the framework to switch its DDMWriter and Reader
to UTF8. Normally, a client would negotiate this through the MGRLVLS
but the ProtocolTest doesn't take this into consideration and has to
be forced into UTF8 mode.

Note that due to DERBY-2031,  each time we add a feature to the JUnit
protocol test we have to do the same for the other or it'll break. 


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@980800 13f79535-47bb-0310-9956-ffa450edef68