DERBY-4757: Implement UTF8 support in DRDA (Client changes)
This patch was contributed by Tiago R. Espinha (tiago dot derby at yahoo dot co dot uk)

This patch lays the foundation for the UTF-8 support in the client:
- Changes the method names for the CcsidManager abstract class so that
they are clearer in terms of what they do (convertToJavaString,
convertFromJavaString) similar to what's been done for the server. 
- Introduces a new name to the CcsidManager that returns, for the given
CcsidManager, the length of a String in bytes (which differs for the
different managers). 
- Creates the Utf8CcsidManager.java for the client and the respective
Utf8CcsidManagerClientTest.java.
- Changes NetAgent.java so that it no longer has a source and target
CCSID managers. Also, similarly to the server, it now has an instance
of each type of CCSID (EBCDIC and UTF-8) and a current CCSID which is
the one currently in use. 

This patch does not yet enable UTF-8 support but puts everything in
place so that the switch mechanism can be implemented. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@982852 13f79535-47bb-0310-9956-ffa450edef68