DERBY-3898: Blob.setBytes differs between embedded and client driver when the specified length is invalid
Added fix and test case for a remaining corner case: When the sum of
offset and length is greater than Integer.MAX_VALUE, the client driver
silently ignores the error whereas the embedded driver fails with an
IndexOutOfBoundsException.

The unexpected results are caused by a check for

  offset + len > bytes.length

where offset+len overflows and evaluates to a negative value. The fix
changes this condition to the equivalent

  len > bytes.length - offset

which won't overflow (because both bytes.length and offset are known
to be non-negative at this point in the code, and subtracting one
non-negative int from another is guaranteed to result in a value in
the range [-Integer.MAX_VALUE, Integer.MAX_VALUE]).


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@986345 13f79535-47bb-0310-9956-ffa450edef68