DERBY-4531 Client setCharacterStream closes its Reader argument stream in finalizer
Patch derby-4531b, which makes the client behave like embedded in this
respect.  

It also adds a new test which shows the clients divergent behavior
prior to the fix in EncodedInputStream. Since the test relies on
explicit gc, it is not guaranteed to show the presence of the bug,
though.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@987331 13f79535-47bb-0310-9956-ffa450edef68