DERBY-4695 Internal assignment of tablenumer, columnnumber looks wrong in query tree, although no ill effects are seen
Patch derby-4695-fixcolumnno-1b, which moves the assignment of
ColumnReference#columnnumber from ColumnReference to the underlying
FromTable, i.e. the same place as where the table number is
assigned. This assures that the two are in synch, and the columnNumber
assignment is no longer thwarted by the (re)assignment of the
ColumnReference#source for "JoinNode"s to one of the result columns of
the Join itself (needed for code generation: the source for the
ColumnReference need to be to the correct ResultSet at generate time,
cf the comment in JoinNode#getMatchingColumn).



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@988204 13f79535-47bb-0310-9956-ffa450edef68