DERBY-1958: Improve XSDG3 error to print container, operation, and name
This patch was contributed by Eranda Sooriyabandara (070468D at gmail dot com)

This patch modifies message XSDG3 ("Metadata for container could not be
accessed") to provide additional information about which container, precisely,
is having the problem. The message now includes the container identity and
the container file name, as well as the type of access to the container
that encountered the problem.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@989696 13f79535-47bb-0310-9956-ffa450edef68