DERBY-3644 and DERBY-2109
As part of DERBY-2109, shutdown command started accepting optional user name and password. Bacause of this change in command, the protcol level was bumped up to 2. But this caused all the network server commands from clients with DERBY-2109 changes to break against a server without the DERBY-2109 changes. That is because the server checks the protocol level of the client commands and if they do match with server protocol version, then server will simply throw an exception about mistmatch of protocol levels and thus causing all the client commands from post DERBY-2109 clients to break against server with pre-DERBY-2109 changes.

Will create another jira to handle the shutdown command's optional user name and password parameters.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@990841 13f79535-47bb-0310-9956-ffa450edef68