DERBY-4786 Shutdown command without username and password should work with mixed client and network server releases.
Commiting changes which modify the really outdated testScript.xml This gets used to run the compaibility tests. In addition to 
trunk, it only tested upto 10.2 and didn't include all releases of 10.2 and 10.1 I have modified it to test all available releases 
upto 10.6 This will definitely increase the runtime of the compatibility tests. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1005193 13f79535-47bb-0310-9956-ffa450edef68