DERBY-4835 Trigger plan does not recompile with upgrade from 10.5.3.0 to 10.6.1.0 causing java.lang.NoSuchMethodError
The problem was that the clearing of stored statements during the upgrade was conditional. The clearing should happen unconditionally during upgrade. Unconditional clearing of SPSes has fixed the problem with this jira. 

Also, added a test in upgrade suite for triggers. The test has been added to BasicSetup.java which will ensure that this trigger test will automatically get run with even future releases. 




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1021426 13f79535-47bb-0310-9956-ffa450edef68