DERBY-4857: Utilize the SOAP API to fetch JIRA issue list for release notes generation
Added the client part of the new functionality.
The client is an independent Maven project, which consists of the code
required to build the RPC infrastructure (based on WSDL and SOAP) and allowing
us to query the JIRA instance used by Apache Derby to generate a list of
issues fixed in a release.
The output of the client is a text file containing information about the
relevant JIRA issues. This file will be fed into the ReleaseNotesGenerator.
The client can perform the following tasks:
 o (primary) generate list of fixed issues
 o (secondary) print release ancestry chain
 o (secondary) print all Derby releases

To build the client manually: cd into 'tools/release/jirasoap', then run
'mvn -Pbuildclient' ('mvn clean' to clean up generated files)
To run the client manually:
    java -jar tools/release/jirasoap/target/JiraSOAP-jar-with-dependencies.jar
(this will print the usage text - read it!)

Patch file: derby-4857-2a-jirasoap_maven_client.diff


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1027056 13f79535-47bb-0310-9956-ffa450edef68