DERBY-4911 restoreIntrFlagIfSeen may throw ShutdownException causing confusing console stack trace at server shutdown
Patch derby-4911, which makes restoreIntrFlagIfSeen ignore
ShutdownException when trying to find lcc. We sometimes saw a
ShutdownException stack dump on the console when stopping the network
server. restoreIntrFlagIfSeen does nothing useful when we are closing
down anyway, so we can safely ignore the exception.

Backported from trunk as svc merge -c 1038813


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.7@1038816 13f79535-47bb-0310-9956-ffa450edef68