DERBY-4834: Add helper class for working with index statistics in JUnit tests
Introduced a timeout value and exepcted count when fetching statistics from
the system table. If a specific number of statistics entries is expected, the
system table will be queried repeatedly until the correct number is obtained or
the query times out. The motivation behind this is that the arrival of
entries in SYS.SYSSTATISTICS isn't that easy to predict since the work is
carried out in a background task.

Also some cleanups in Utilities:
 o fixed some JavaDoc
 o fixed formatting
 o added method sleep
 o simplified repeatChar

Patch file: derby-4834-3a-timeout.diff



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1041374 13f79535-47bb-0310-9956-ffa450edef68