DERBY-4920 suites.All stuck in RAFContainer4.awaitRestoreChannel()
Patch derby-4920-a. 

Patch details:

* InterruptStatus.java

Make sure we always catch ShutdownException when we access the context
looking for lcc.

* RAFContainer4.java

Improved some comments, added a final, added a max count for a retry
loop that missed it: will throw FILE_IO_INTERRUPTED if it exceeds
MAX_INTERRUPT_RETRIES (as elsewhere when we give up).



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1049150 13f79535-47bb-0310-9956-ffa450edef68