DERBY-4157 backport refactor of upgrade tests. Contributed by Rick Hillegas
Backporting change 764912 which is just a code refactor. The remaining DERBY-4157 have not been ported to 10.5.

Attaching derby-4157-01-aa-refactor.diff. This patch is a first step toward writing the new test. This patch factors out some of the existing upgrade test machinery so that the new test can re-use that machinery:
+
1) logic to discover old versions
2) logic to create class loaders out of those old versions
3) logic to use those class loaders




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.5@1055585 13f79535-47bb-0310-9956-ffa450edef68