DERBY-4980 Replication tests lose derby.log for slave
Patch derby-4980-1 which fixes this. 

The error is that when starting the Derby server for the slave, the
same path name is used for the slave server's working directory as for
the master server ("db_master"). This means that the slave server
doesn't get its own log file.

The patch sets the working directory for the slave correctly to the
"db_slave" directory.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1067366 13f79535-47bb-0310-9956-ffa450edef68