DERBY-5002 In case of checksum error, insure correct error reported to user.
backported change #1067357 from trunk to 10.7 branch.

Changed the order of sanity page checking to make sure that if there is a
checksum error on reading the page from disk, then it is the error reported.
Before this change certain page inconsistencies would be found before doing
the checksum check and would report different kinds of errors depending on
where the corruption happened on the page. The main error case checksums
try to catch is partially written page where because a derby page is made up
of multiple OS/drive blocks some blocks can make it to disk before others and
in the case of a hardware crash an incomplete page may be written. In this
case, the current derby implementation can not recover from log as it needs
a valid page to look at in order to apply log records. The db must be
recovered from derby backup in this case. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.7@1068273 13f79535-47bb-0310-9956-ffa450edef68