DERBY-4940: Tune istat operational characteristics DERBY-5038: updatelocks test failing in tinderbox when background update statistics are enabled.
Changed the default absdiff threshold from zero to 1000, as the current setting
causes the automatic index statistics update feature to be too aggressive.
It also causes a set of regression tests to fail (not all of them fails
always).
The "diff mechanism" may be replaced later, for instance with a plain lower
limit on the number of rows in the table, discussion is ongoing.

Patch file: derby-4940-1a-absdiff_to_1k.diff



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1071588 13f79535-47bb-0310-9956-ffa450edef68