DERBY-4741 Make embedded Derby work reliably in the presence of thread interrupts
Patch derby-4741-sleeps-waits-more, which "regularizes" a few more
instances of interrrupt handling to follow the idiom established in
this issue's patches. 

This leaves a few instances in BasicDaemon.java (as far as embedded
code is concerned), which will need more consideration. In any case,
interrupting the demon threads is less of a valid use case I believe,
i.e. Derby's ability to tolerate that is less crucial that tolerating
interrupts to the user's connection threads.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1073595 13f79535-47bb-0310-9956-ffa450edef68