DERBY-5069 Since Feb 7,2011 weme 6.2 Junit tests have failed to run completely with Failed to invoke suite():java.lang.reflect.InvocationTargetException
Prevent Driver tests from running under J2ME which were preventing suites.All from running. The problem was introduced with revision 1067954 with some additions to DriverTest. It is not really clear when  JDBC.vmSupportsJSR169() needs to be called in the the individual test vs the suite calling the test.  In order to get the tests running again, this change moves the DriverTest check for JSR169 to the suite.  Also added stack traces to the output when an invocation error occurs to make it easier to track down.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1074762 13f79535-47bb-0310-9956-ffa450edef68