DERBY-4319 hang in suites.all with ibm 1.5 on AIX after ttestDefaultProperties
Make sure spawned network server process is destroyed if the hang occurs on SpawnedProcess.complete(). This does not address the core reason for the hang which is still not well understood and currently does not reproduce on the aix machine with the problem. DERBY-4319 will stay open while we wait for the problem to reappear.





git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1079548 13f79535-47bb-0310-9956-ffa450edef68