DERBY-5108
Changes istat daemon shutdown to check during processing if a shutdown is
in progress and respond to the shutdown immediately.  Also changes the
module stop() to wait for worker threads to exit before returning.  Waiting
for work to stop allows the subsequent shutdown of the storage system to
properly close it's files during a clean shutdown request.  Without this
change the system sometimes left files open which the nightly tests uncovered
on windows machines while trying to delete those files.

This change is a slightly modified version of a patch proposed by Knut Anders Ha
tlen.

Previous to this change the AutomaticIndexStatisticsTest.testShutdownWhileScanni
ngThenDelete
test would fail on my machine consistently in SANE classes mode on a windows XP
laptop.  After this
I have only seen one failure in 50 runs.  Checking it in as it is definitely
an improvement and want to see if it fixes the errors in the nightly's across
a number of environments.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1081677 13f79535-47bb-0310-9956-ffa450edef68