DERBY-5121 Data corruption when executing an UPDATE trigger
DERBY-5121 corruption is caused by the changes that went in for DERBY-1482. DERBY-1482 changes went into 10.7 codeline and trunk. 
There is already a release out off of 10.7 with DERBY-1482 changes and that release is 10.7.1.1. We haven't had a release off of
trunk codeline yet and hence there are no corrupt databases based of trunk codelines but there can be corrupt databases based
off 10.7.1.1 release.

The fix for courrption problem is to back out DERBY-1482 changes. Since 10.7.1.1, We have backed out DERBY-1482 changes from 10.7
and trunk codeline and this fix will be available in the next subsequent releases for those 2 codelines. 

As for the existing 10,7.1.1 users, they need to make sure that their 10.7.1.1 databases get upgraded to next release with 
DERBY-1482 changes backed out. As part of the upgrades, the corrupt trigger action SPSes will get marked invalid and when they get 
fired next time around, correct trigger action SPSes will be generated for them,

In order to provide this upgrade path to fix corrupt triggers, we need to increment the point release number for 10.7.

With this checkin, I am bumping the release number for next 10.7 release. This will ensure that the upgrade path will be followed
and the trigger action SPSes will get fixed because now we have backed out DERBY-1482 changes.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.7@1086418 13f79535-47bb-0310-9956-ffa450edef68