DERBY-5152 Shutting down db, information that the thread received an interrupt will not be restored to thread's interrupt flag
Patch derby-5152-b. When a thread receives an interrupt Derby detects
this, it will reset the thread's flag and save the fact in its lcc
(LanguageConnectionContext), if available. If not (e.g. during boot)
it will save the information in a thread local variable. For
performance reasons, we use the lcc when available. However, when
shutting down the database, the lcc goes away, and when the JDBC call
returns to the application, the thread's interrupt flag will not be
reinstated as per our specification. This is because the lcc dies
before we do the restoring (under shutdown). So, the information that
the thread was interrupted is lost with the lcc going away.

This patch copies the information from lcc over to the thread local
variable when lcc is popped and adds a new test case to
InterruptResilienceTest.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1086443 13f79535-47bb-0310-9956-ffa450edef68