DERBY-5185 store/rollForwardRecovery.sql stuck in RAFContainer4.recoverContainerAfterInterrupt() during shutdown
Patch derby-5185-1a.diff. Avoid waiting forever in the loop in
recoverContainerAfterInterrupt where we wait for other concurrent
threads to hit the wall (having seen ClosedChannelException), i.e. so
we know they waiting for this thread to clean up. The counting logic
(threadsInPageIO) here needs to be correct, if there is an error, we
could risk waiting forever, as seen in this issue.

This patch should be followed up by a patch to correct the logic, but
until such time, this patch improves on the situation.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1091221 13f79535-47bb-0310-9956-ffa450edef68