DERBY-2354: Unable to perform select query using DISTINCT on a read-only database
Made databases that are stored in a jar or a zip use a writable
directory for temporary files needed by some operations (typically
hash scans when the hash table gets too big to be kept in memory).

The writable directory can be controlled by the database property
derby.storage.tempDirectory.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1092067 13f79535-47bb-0310-9956-ffa450edef68