DERBY-4260: Make derbynet/NetworkServerControlClientCommandTest run regardless of the locale
This patch was contributed by Houx Zhang (houxzhang @ gmail dot com).

This patch modifies the NetworkServerControlClientCommandTest to pass

	-Dderby.ui.locale=en_US

when running the sub-commands of the test, to ensure that the test assertions,
which look for specific English language strings, find the expected values.
That way, the test will pass, even if run in a shell environment where the
locale is set to a non-English value.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1095247 13f79535-47bb-0310-9956-ffa450edef68