DERBY-4443: Wrap rollback in exception handlers in try-catch
This patch was contributed by Houx Zhang (houxzhang at gmail dot com)

The patch modifies the error handling in SystemProcedures.java so that
secondary exceptions during rollback are chained to the primary exception.
Also, EmbedSQLException's wrapping behavior is adjusted so that the
StandardException logic doesn't try to unwrap the chained exceptions.

A new test is added to the tools suite; it simulates rollback errors
using a special mock driver and mock connection, and verifies that
the exceptions are chained properly.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1096991 13f79535-47bb-0310-9956-ffa450edef68